requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
setuptools==75.2.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
